requirements.txt
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-ollama>=0.1.0 
boto3>=1.34.0
opensearch-py>=2.4.1
requests-aws4auth>=1.2.3
python-dotenv>=1.0.0
unstructured>=0.10.30
faiss-cpu>=1.7.4  # Use faiss-gpu instead if you want GPU support 
